LRN: Use regexps in logdefs
Now "*" is not valid anymore (it's not like anyone had a lot of opportunities
to use it though).
Empty string means ".*", so it's still compatible.
Matching is case-sensitive. ^ and $ match the beginning and the end of
the text, newine matches a newline character.
All regexps are considered non-extended.
Regex compilation errors are not reported in any way (the parsing process just
fails, and that's it).



git-svn-id: https://gnunet.org/svn/gnunet@18016 140774ce-b5e7-0310-ab8b-a85725594a96