Applying 1st patch from Andrew Cann (see also #2887).
https://canndrew.org/misc/trim_abstract_socket_paths.diff [^]

This changes the addrlen argument passed to the connect, bind and sendto
syscalls to reflect the length of the string in sockaddr_un.sun_path. I
was trying to talk to GNUnet services using socat and discovered that it
expects abstract sockets to be created this way and won't connect
otherwise. Looking at some other programs that I could talk to using
socat (dbus, X11), this is how they create abstract sockets as well.


git-svn-id: https://gnunet.org/svn/gnunet@29635 140774ce-b5e7-0310-ab8b-a85725594a96